Moved analysis package to directory hqAnalysis
In order to separate puzzle scripts from the analysis
package I crated a hqAnalysis folder, moved the
package hqhelp there and then added the empty
__init__.py to ensure that python reads the new
directory for packages.  This required that I update the
import in hq-1-1.py
