+++ cirq/HelloQuantum/hqAnalysis/hqhelp.py
